requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.4
Deprecated==1.2.14
Django==4.2.5
django-oauth-toolkit==2.3.0
django-stdimage==6.0.1
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
Pillow==10.0.1
psycopg2==2.9.7
pycparser==2.21
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.5
wrapt==1.15.0
